Creat models,controllers to store the list for the players
