+++ src/main/kotlin/controllers/PlayersAPI.java
@@ -1,4 +1,12 @@
-package controllers;
+package controllers
 
-public class PlayersAPI {
+import models.Players
+import persistence.Serializer
+
+class PlayersAPI(serializerType: Serializer) {
+
+    private var serializer: Serializer = serializerType
+
+    private var players = ArrayList<Players>()
+
 }
