--- src/main/kotlin/models/Players.java
@@ -1,11 +1,0 @@
-package models
-
-data class Players(
-        var playerName: String,
-        var playerSurname: String,
-        var age: Int,
-        var height: Int,
-        var weight: Int,
-        var position: String,
-        var isNoteArchived: Boolean)
-{}
